Puntjes op de i en alles werkend gekregen
Input kan meegegeven worden op de commandline.
Spaties tussen waarden worden vanzelf verwijderd tussen het uitlezen.
Tasks aan het einde van Jobs worden nu wel gesloten.
Tasks uit beïndigde Jobs worden niet herhaaldelijk gesloten.

+++ Task.h
@@ -2,7 +2,7 @@
  * Task.h
  *
  *  Created on: 12 Sep 2021
- *      Author: Damo Luijpers
+ *      Author: Damo Luijpers, Rick Smallenbroek
  */
 
 #ifndef TASK_H_
@@ -20,8 +20,8 @@ public:
 	//Getters & Setters
 	const unsigned short getDuration() const;
 	unsigned short getMachineId() const;
-	unsigned short getEndTime() const;
-	void setEndTime(unsigned short startTime);
+	unsigned long long getEndTime() const;
+	void setEndTime(unsigned long long startTime);
 	bool isTaskCompleted() const;
 	void setTaskCompleted(bool taskCompleted);
 
@@ -29,7 +29,7 @@ private:
 	//Attributes
 	unsigned short machineId;
 	unsigned short duration;
-	unsigned short endTime;
+	unsigned long long endTime;
 	bool taskCompleted;
 };
